cloud_run_job/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-cloud-bigquery==3.11.4
google-auth==2.23.4